fucking last patch for x86_64 compilation
git-svn-id: https://piggledy.org/svn/peerfuse/trunk@929 f7b84f85-2e04-4de5-9e99-c4612dc0539c

+++ common/packet_base.cpp
@@ -257,10 +257,10 @@ PacketBase& PacketBase::Write(FileChunk chunk)
 PacketBase& PacketBase::Write(const Certificate& cert)
 {
 	unsigned char *buf;
-	uint32_t str_len;
+	size_t str_len;
 	cert.GetRaw(&buf, &str_len);
 
-	Write(str_len);
+	Write((uint32_t)str_len);
 	char* new_datas = new char [size + str_len];
 	if(datas)
 		memcpy(new_datas, datas, size);
